+++ Utils/src/main/java/com/distrimind/util/crypto/SymmetricAuthenticatedSignatureCheckerAlgorithm.java
@@ -35,6 +35,8 @@ knowledge of the CeCILL-C license and that you accept its terms.
 package com.distrimind.util.crypto;
 
 
+import com.distrimind.bouncycastle.util.Arrays;
+
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -99,12 +101,7 @@ public class SymmetricAuthenticatedSignatureCheckerAlgorithm extends AbstractAut
 			if (signature == null)
 				return false;
 			byte[] mySignature = signer.getSignature();
-			if (mySignature.length != signature.length)
-				return false;
-			for (int i = 0; i < mySignature.length; i++)
-				if (mySignature[i] != signature[i])
-					return false;
-			return true;
+			return Arrays.constantTimeAreEqual(mySignature, signature);
 		} catch (IOException e) {
 			return false;
 		} finally